Reformat code and added some documentation

+++ src/main/java/dev/heliosclient/mixin/CapeFeatureRendererMixin.java
@@ -1,7 +1,7 @@
 package dev.heliosclient.mixin;
 
-import dev.heliosclient.managers.ModuleManager;
 import dev.heliosclient.managers.CapeManager;
+import dev.heliosclient.managers.ModuleManager;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
@@ -17,7 +17,8 @@ import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.math.*;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.RotationAxis;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -30,6 +31,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class CapeFeatureRendererMixin extends FeatureRenderer<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> {
     @Unique
     private int ticksInWater = 0;
+
     public CapeFeatureRendererMixin(FeatureRendererContext<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> context) {
         super(context);
     }
@@ -38,20 +40,20 @@ public abstract class CapeFeatureRendererMixin extends FeatureRenderer<AbstractC
     public void render(MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, AbstractClientPlayerEntity abstractClientPlayerEntity, float f, float g, float h, float j, float k, float l, CallbackInfo ci) {
 
         if (!abstractClientPlayerEntity.isInvisible() && abstractClientPlayerEntity.isPartVisible(PlayerModelPart.CAPE) && ModuleManager.capeModule.isActive()) {
-            if(CapeManager.shouldPlayerHaveCape(abstractClientPlayerEntity)){
+            if (CapeManager.shouldPlayerHaveCape(abstractClientPlayerEntity)) {
                 Identifier capeTexture = CapeManager.getCapeTexture(abstractClientPlayerEntity);
                 ItemStack itemStack = abstractClientPlayerEntity.getEquippedStack(EquipmentSlot.CHEST);
                 if (!itemStack.isOf(Items.ELYTRA)) {
                     matrixStack.push();
                     matrixStack.translate(0.0F, 0.0F, 0.125F);
 
-                    double d = MathHelper.lerp((double) h, abstractClientPlayerEntity.prevCapeX, abstractClientPlayerEntity.capeX) - MathHelper.lerp((double) h, abstractClientPlayerEntity.prevX, abstractClientPlayerEntity.getX());
-                    double e = MathHelper.lerp((double) h, abstractClientPlayerEntity.prevCapeY, abstractClientPlayerEntity.capeY) - MathHelper.lerp((double) h, abstractClientPlayerEntity.prevY, abstractClientPlayerEntity.getY());
-                    double m = MathHelper.lerp((double) h, abstractClientPlayerEntity.prevCapeZ, abstractClientPlayerEntity.capeZ) - MathHelper.lerp((double) h, abstractClientPlayerEntity.prevZ, abstractClientPlayerEntity.getZ());
+                    double d = MathHelper.lerp(h, abstractClientPlayerEntity.prevCapeX, abstractClientPlayerEntity.capeX) - MathHelper.lerp(h, abstractClientPlayerEntity.prevX, abstractClientPlayerEntity.getX());
+                    double e = MathHelper.lerp(h, abstractClientPlayerEntity.prevCapeY, abstractClientPlayerEntity.capeY) - MathHelper.lerp(h, abstractClientPlayerEntity.prevY, abstractClientPlayerEntity.getY());
+                    double m = MathHelper.lerp(h, abstractClientPlayerEntity.prevCapeZ, abstractClientPlayerEntity.capeZ) - MathHelper.lerp(h, abstractClientPlayerEntity.prevZ, abstractClientPlayerEntity.getZ());
                     float n = MathHelper.lerpAngleDegrees(h, abstractClientPlayerEntity.prevBodyYaw, abstractClientPlayerEntity.bodyYaw);
 
-                    double o = (double) MathHelper.sin(n * 0.017453292F);
-                    double p = (double) (-MathHelper.cos(n * 0.017453292F));
+                    double o = MathHelper.sin(n * 0.017453292F);
+                    double p = -MathHelper.cos(n * 0.017453292F);
 
                     float q = (float) e * 10.0F;
                     q = MathHelper.clamp(q, -6.0F, 32.0F);
@@ -76,10 +78,9 @@ public abstract class CapeFeatureRendererMixin extends FeatureRenderer<AbstractC
                     if (ModuleManager.capeModule.customPhysics.value) {
                         //New physics
                         if (abstractClientPlayerEntity.isSubmergedInWater()) {
-                            if(ticksInWater >= 800){
+                            if (ticksInWater >= 800) {
                                 ticksInWater = 800;
-                            }
-                            else{
+                            } else {
                                 ticksInWater++;
                             }
                             // Adjust the cape's rotation to make it float upwards
@@ -98,7 +99,6 @@ public abstract class CapeFeatureRendererMixin extends FeatureRenderer<AbstractC
                     }
 
 
-
                     matrixStack.multiply(RotationAxis.POSITIVE_Z.rotationDegrees(s / 2.0F));
                     matrixStack.multiply(RotationAxis.POSITIVE_Y.rotationDegrees(180.0F - s / 2.0F));
 
@@ -110,6 +110,7 @@ public abstract class CapeFeatureRendererMixin extends FeatureRenderer<AbstractC
             }
         }
     }
+
     @ModifyVariable(method = "render*", at = @At("STORE"), ordinal = 6)
     private float render(float n, MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, AbstractClientPlayerEntity abstractClientPlayerEntity, float f, float g, float h, float j, float k, float l) {
         return MathHelper.lerp(h, abstractClientPlayerEntity.prevBodyYaw, abstractClientPlayerEntity.bodyYaw);
